+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Render links inline when the text is the same as the path.
 - Handle edge-case where code block is not contained in a `p` tag.
 
 ## 0.1.1